[eas-cli] replace API newAndroid -> android
