--- packages/eas-cli/src/credentials/__mocks__/context.ts
@@ -1,7 +1,0 @@
-import { getAndroidApiMock } from '../__tests__/fixtures-android';
-import { createCtxMock } from '../__tests__/fixtures-context';
-
-export const createCredentialsContextAsync = () =>
-  createCtxMock({
-    android: getAndroidApiMock(),
-  });
