+++ packages/eas-cli/src/credentials/__tests__/fixtures-context.ts
@@ -1,8 +1,7 @@
 import merge from 'lodash/merge';
 
 import { Context } from '../context';
-import { getAndroidApiMockWithoutCredentials } from './fixtures-android';
-import { getNewAndroidApiMockWithoutCredentials } from './fixtures-android-new';
+import { getNewAndroidApiMockWithoutCredentials } from './fixtures-android';
 import { getAppstoreMock } from './fixtures-appstore';
 import { testAppJson, testUsername } from './fixtures-constants';
 import { getNewIosApiMockWithoutCredentials } from './fixtures-ios';
@@ -10,8 +9,7 @@ import { getNewIosApiMockWithoutCredentials } from './fixtures-ios';
 export function createCtxMock(mockOverride: Record<string, any> = {}): Context {
   const defaultMock = {
     ios: getNewIosApiMockWithoutCredentials(),
-    android: getAndroidApiMockWithoutCredentials(),
-    newAndroid: getNewAndroidApiMockWithoutCredentials(),
+    android: getNewAndroidApiMockWithoutCredentials(),
     appStore: getAppstoreMock(),
     bestEffortAppStoreAuthenticateAsync: jest.fn(),
     ensureAppleCtx: jest.fn(),
