+++ packages/eas-cli/src/credentials/android/actions/AssignFcm.ts
@@ -13,10 +13,10 @@ export class AssignFcm {
     ctx: Context,
     fcm: AndroidFcmFragment
   ): Promise<CommonAndroidAppCredentialsFragment> {
-    const appCredentials = await ctx.newAndroid.createOrGetExistingAndroidAppCredentialsWithBuildCredentialsAsync(
+    const appCredentials = await ctx.android.createOrGetExistingAndroidAppCredentialsWithBuildCredentialsAsync(
       this.app
     );
-    const updatedAppCredentials = await ctx.newAndroid.updateAndroidAppCredentialsAsync(
+    const updatedAppCredentials = await ctx.android.updateAndroidAppCredentialsAsync(
       appCredentials,
       {
         androidFcmId: fcm.id,
