+++ packages/eas-cli/src/credentials/android/actions/__tests__/AssignFcm-test.ts
@@ -1,4 +1,4 @@
-import { testLegacyAndroidFcmFragment } from '../../../__tests__/fixtures-android-new';
+import { testLegacyAndroidFcmFragment } from '../../../__tests__/fixtures-android';
 import { createCtxMock } from '../../../__tests__/fixtures-context';
 import { AssignFcm } from '../AssignFcm';
 import { getAppLookupParamsFromContext } from '../BuildCredentialsUtils';
@@ -14,9 +14,9 @@ describe(AssignFcm, () => {
 
     // expect app credentials to be fetched/created, then updated
     expect(
-      ctx.newAndroid.createOrGetExistingAndroidAppCredentialsWithBuildCredentialsAsync as any
+      ctx.android.createOrGetExistingAndroidAppCredentialsWithBuildCredentialsAsync as any
     ).toHaveBeenCalledTimes(1);
-    expect(ctx.newAndroid.updateAndroidAppCredentialsAsync as any).toHaveBeenCalledTimes(1);
+    expect(ctx.android.updateAndroidAppCredentialsAsync as any).toHaveBeenCalledTimes(1);
   });
   it('works in Non-Interactive Mode', async () => {
     const ctx = createCtxMock({
