+++ packages/eas-cli/src/credentials/android/actions/__tests__/RemoveFcm-test.ts
@@ -2,7 +2,7 @@ import { confirmAsync } from '../../../../prompts';
 import {
   getNewAndroidApiMockWithoutCredentials,
   testAndroidAppCredentialsFragment,
-} from '../../../__tests__/fixtures-android-new';
+} from '../../../__tests__/fixtures-android';
 import { createCtxMock } from '../../../__tests__/fixtures-context';
 import { getAppLookupParamsFromContext } from '../BuildCredentialsUtils';
 import { RemoveFcm } from '../RemoveFcm';
@@ -24,7 +24,7 @@ describe(RemoveFcm, () => {
   it('removes an FCM Api Key', async () => {
     const ctx = createCtxMock({
       nonInteractive: false,
-      newAndroid: {
+      android: {
         ...getNewAndroidApiMockWithoutCredentials(),
         getAndroidAppCredentialsWithCommonFieldsAsync: jest.fn(
           () => testAndroidAppCredentialsFragment
@@ -34,7 +34,7 @@ describe(RemoveFcm, () => {
     const appLookupParams = getAppLookupParamsFromContext(ctx);
     const removeFcmApiKeyAction = new RemoveFcm(appLookupParams);
     await removeFcmApiKeyAction.runAsync(ctx);
-    expect(ctx.newAndroid.deleteFcmAsync as any).toHaveBeenCalledTimes(1);
+    expect(ctx.android.deleteFcmAsync as any).toHaveBeenCalledTimes(1);
   });
   it('errors in Non-Interactive Mode', async () => {
     const ctx = createCtxMock({ nonInteractive: true });
