+++ packages/eas-cli/src/credentials/android/actions/__tests__/SetupBuildCredentials-test.ts
@@ -3,7 +3,7 @@ import {
   getNewAndroidApiMockWithoutCredentials,
   testAndroidBuildCredentialsFragment,
   testJksAndroidKeystoreFragment,
-} from '../../../__tests__/fixtures-android-new';
+} from '../../../__tests__/fixtures-android';
 import { createCtxMock } from '../../../__tests__/fixtures-context';
 import { MissingCredentialsNonInteractiveError } from '../../../errors';
 import { getAppLookupParamsFromContext } from '../BuildCredentialsUtils';
@@ -27,7 +27,7 @@ describe('SetupBuildCredentials', () => {
   it('skips setup when there are prior credentials', async () => {
     const ctx = createCtxMock({
       nonInteractive: false,
-      newAndroid: {
+      android: {
         ...getNewAndroidApiMockWithoutCredentials(),
         getDefaultAndroidAppBuildCredentialsAsync: jest.fn(
           () => testAndroidBuildCredentialsFragment
@@ -39,13 +39,13 @@ describe('SetupBuildCredentials', () => {
     await setupBuildCredentialsAction.runAsync(ctx);
 
     // expect keystore not to be created
-    expect(ctx.newAndroid.createKeystoreAsync as any).toHaveBeenCalledTimes(0);
+    expect(ctx.android.createKeystoreAsync as any).toHaveBeenCalledTimes(0);
   });
   it('sets up credentials when there are no prior credentials', async () => {
     (promptAsync as jest.Mock).mockImplementation(() => ({ providedName: 'test-provided-name' }));
     const ctx = createCtxMock({
       nonInteractive: false,
-      newAndroid: {
+      android: {
         ...getNewAndroidApiMockWithoutCredentials(),
         createKeystoreAsync: jest.fn(() => testJksAndroidKeystoreFragment),
       },
@@ -55,13 +55,13 @@ describe('SetupBuildCredentials', () => {
     await setupBuildCredentialsAction.runAsync(ctx);
 
     // expect keystore to be created
-    expect(ctx.newAndroid.createKeystoreAsync as any).toHaveBeenCalledTimes(1);
+    expect(ctx.android.createKeystoreAsync as any).toHaveBeenCalledTimes(1);
   });
   it('errors in Non-Interactive Mode', async () => {
     (promptAsync as jest.Mock).mockImplementation(() => ({ providedName: 'test-provided-name' }));
     const ctx = createCtxMock({
       nonInteractive: true,
-      newAndroid: {
+      android: {
         ...getNewAndroidApiMockWithoutCredentials(),
       },
     });
